moving repeating code to one location
