+++ components/pages/submission-system/program-management/index.tsx
@@ -1,30 +1,14 @@
 import React from 'react';
-import useAuthContext from 'global/hooks/useAuthContext';
 import { useRouter } from 'next/router';
 import { css } from 'uikit';
-import { ContentBox } from 'uikit/PageLayout';
 import TitleBar from 'uikit/TitleBar';
 import SubmissionLayout from '../layout';
 import ManageProgramTabs from './ManageProgramTabs';
-import { useQuery } from '@apollo/react-hooks';
-import { GqlClinicalSubmissionData } from '../program-clinical-submission/types';
-import SIDE_MENU_CLINICAL_SUBMISSION_STATE from '../SIDE_MENU_CLINICAL_SUBMISSION_STATE.gql';
-import Notification from 'uikit/notifications/Notification';
+import { SchemaInvalidSubmisisonNotification } from '../SchemaInvalidSubmissionNotification';
 
 export default () => {
   const router = useRouter();
   const { shortName: programShortName } = router.query;
-
-  const { data: { clinicalSubmissions = undefined } = {} } = useQuery<{
-    clinicalSubmissions: GqlClinicalSubmissionData;
-  }>(SIDE_MENU_CLINICAL_SUBMISSION_STATE, {
-    variables: {
-      programShortName: programShortName,
-    },
-  });
-
-  const hasSchemaErrorsAfterMigration =
-    clinicalSubmissions && clinicalSubmissions.state === 'INVALID_BY_MIGRATION';
 
   return (
     <SubmissionLayout
@@ -43,18 +27,12 @@ export default () => {
         </div>
       }
     >
-      {hasSchemaErrorsAfterMigration && (
-        <Notification
-          css={css`
-            margin-bottom: 20px;
-          `}
-          size="SM"
-          variant="ERROR"
-          title={`Your clinical submission is invalid`}
-          content={`Version _link here_ was released and has made your clinical submission invalid. See the details in your clinical workspace.`}
-          interactionType="ACTION_DISMISS"
+      {
+        <SchemaInvalidSubmisisonNotification
+          marginBottom={20}
+          programShortName={programShortName as string}
         />
-      )}
+      }
       <ManageProgramTabs />
     </SubmissionLayout>
   );
