+++ components/pages/submission-system/program-sample-registration/index.tsx
@@ -7,14 +7,12 @@ import { Row } from 'react-grid-system';
 import { css } from 'uikit';
 import Button, { BUTTON_SIZES, BUTTON_VARIANTS } from 'uikit/Button';
 import Container from 'uikit/Container';
-import DnaLoader from 'uikit/DnaLoader';
 import Link from 'uikit/Link';
 import Progress, { PROGRESS_STATUS } from 'uikit/Progress';
 import TitleBar from 'uikit/TitleBar';
 import Typography from 'uikit/Typography';
 import SubmissionLayout from '../layout';
 import CLEAR_CLINICAL_REGISTRATION_MUTATION from './gql/CLEAR_CLINICAL_REGISTRATION_MUTATION.gql';
-import SIDE_MENU_CLINICAL_SUBMISSION_STATE from '../SIDE_MENU_CLINICAL_SUBMISSION_STATE.gql';
 import FileTable from './FileTable';
 import NoDataMessage from './FileTable/NoDataMessage';
 import GET_REGISTRATION from './gql/GET_REGISTRATION.gql';
@@ -24,11 +22,9 @@ import { containerStyle } from '../common';
 import { useToaster } from 'global/hooks/toaster';
 import ErrorNotification, { defaultColumns } from '../ErrorNotification';
 import { ClinicalRegistrationData, ClinicalRegistration } from './types';
-import Notification, { NOTIFICATION_INTERACTION_EVENTS } from 'uikit/notifications/Notification';
+import Notification from 'uikit/notifications/Notification';
 import { toDisplayError } from 'global/utils/clinicalUtils';
-import { GqlClinicalSubmissionData } from '../program-clinical-submission/types';
-import { useRouter } from 'next/router';
-import { PROGRAM_CLINICAL_SUBMISSION_PATH, PROGRAM_SHORT_NAME_PATH } from 'global/constants/pages';
+import { SchemaInvalidSubmisisonNotification } from '../SchemaInvalidSubmissionNotification';
 
 const recordsToFileTable = (
   records: ClinicalRegistrationData[],
@@ -56,19 +52,6 @@ export default function ProgramIDRegistration() {
     variables: { shortName: programShortName },
   });
 
-  const { data: { clinicalSubmissions = undefined } = {} } = useQuery<{
-    clinicalSubmissions: GqlClinicalSubmissionData;
-  }>(SIDE_MENU_CLINICAL_SUBMISSION_STATE, {
-    variables: {
-      programShortName: programShortName,
-    },
-  });
-
-  const hasSchemaErrorsAfterMigration =
-    clinicalSubmissions && clinicalSubmissions.state === 'INVALID_BY_MIGRATION';
-
-  const [closedMigrationMsg, setclosedMigrationMsg] = React.useState(false);
-
   const schemaOrValidationErrors = get(
     clinicalRegistration,
     'errors',
@@ -93,7 +76,6 @@ export default function ProgramIDRegistration() {
   const [clearRegistration] = useMutation(CLEAR_CLINICAL_REGISTRATION_MUTATION);
 
   const toaster = useToaster();
-  const router = useRouter();
 
   const handleClearClick = async () => {
     if (clinicalRegistration.id == null) {
@@ -215,27 +197,7 @@ export default function ProgramIDRegistration() {
         </div>
       }
     >
-      {hasSchemaErrorsAfterMigration && !closedMigrationMsg && (
-        <Notification
-          size="SM"
-          variant="ERROR"
-          title={`Your clinical submission is invalid`}
-          content={`Version _link here_ was released and has made your clinical submission invalid. See the details in your clinical workspace.`}
-          interactionType="ACTION_DISMISS"
-          onInteraction={({ type }) => {
-            if (type === NOTIFICATION_INTERACTION_EVENTS.ACTION) {
-              router.push(
-                PROGRAM_CLINICAL_SUBMISSION_PATH.replace(
-                  PROGRAM_SHORT_NAME_PATH,
-                  programShortName as string,
-                ),
-              );
-            } else {
-              setclosedMigrationMsg(true);
-            }
-          }}
-        />
-      )}
+      {<SchemaInvalidSubmisisonNotification programShortName={programShortName as string} />}
       <Container
         css={css`
           ${containerStyle}
